fixed a few bugs with matches and conversations

+++ app/controllers/matches_controller.rb
@@ -12,8 +12,12 @@ class MatchesController < ApplicationController
     user2 = User.find(match.matcher_id)
     match1 = Match.where(matcher_id: user1.id,matched_id: user2.id)[0]
     match2 = Match.where(matcher_id: user2.id,matched_id: user1.id)[0]
-    match1.destroy
-    match2.destroy
+    if match1
+      match1.destroy
+    end
+    if match2
+      match2.destroy
+    end
     redirect_back(fallback_location: current_user)
   end
 end
